requirements.txt
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
Babel==2.10.3
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.1
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap-v5==1.0.11
django-bootstrap3==22.1
django-bootstrap4==22.2
django-crispy-forms==1.14.0
djangorestframework==3.13.1
frozenlist==1.3.1
idna==3.3
multidict==6.0.2
Pillow==9.2.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
yarl==1.8.1
